Fixed SSE int32/int64 type in libdivide_sum_* horizontal add helpers

+++ libdivide_benchmark.c
@@ -109,11 +109,11 @@ uint32_t libdivide_sum_4u32(__m128i x) {
 	const uint32_t *comps = (const uint32_t*)&x;
 	return comps[0] + comps[1] + comps[2] + comps[3];
 }
-int32_t libdivide_sum_2s64(__m128i x) {
+int64_t libdivide_sum_2s64(__m128i x) {
 	const int64_t *comps = (const int64_t*)&x;
 	return comps[0] + comps[1];
 }
-int32_t libdivide_sum_2u64(__m128i x) {
+uint64_t libdivide_sum_2u64(__m128i x) {
 	const uint64_t *comps = (const uint64_t*)&x;
 	return comps[0] + comps[1];
 }
